now using context in miner; closes #57

+++ core/chain/mainchain/miner/service.go
@@ -112,8 +112,9 @@ func (s Service) buildMainchainBlockAsync() error {
 		wg.Add(1)
 		go func(iHash string, txs []*statechain.Transaction) {
 			defer wg.Done()
-			if s.props.IsValid == nil || *s.props.IsValid == false {
-				err = errors.New("miner is invalid")
+			if s.props.Context.Err() != nil {
+				err = s.props.Context.Err()
+
 				return
 			}
 
@@ -143,8 +144,8 @@ func (s Service) buildMainchainBlock() error {
 
 	// 2. apply txs
 	for imageHash, transactions := range txsMap {
-		if s.props.IsValid == nil || *s.props.IsValid == false {
-			return errors.New("miner is invalid")
+		if s.props.Context.Err() != nil {
+			return s.props.Context.Err()
 		}
 
 		if err := s.buildNextStates(imageHash, transactions); err != nil {
@@ -164,8 +165,8 @@ func (s Service) mineBlock() error {
 	}
 
 	for {
-		if s.props.IsValid == nil || *s.props.IsValid == false {
-			return errors.New("miner is invalid")
+		if s.props.Context.Err() != nil {
+			return s.props.Context.Err()
 		}
 
 		hash, nonce, err := s.generateHashAndNonce()
@@ -178,14 +179,16 @@ func (s Service) mineBlock() error {
 			return err
 		}
 
-		nextProps := s.minedBlock.NextBlock.Props()
-		nextProps.Nonce = nonce
-		nextBlock := mainchain.New(&nextProps)
-		s.minedBlock.NextBlock = nextBlock
-
 		if check {
+			nextProps := s.minedBlock.NextBlock.Props()
+			nextProps.Nonce = nonce
+			nextBlock := mainchain.New(&nextProps)
+			s.minedBlock.NextBlock = nextBlock
+
 			return s.minedBlock.NextBlock.SetHash()
 		}
+
+		// note: else the for loop continues and we try the next hash
 	}
 }
 
@@ -198,6 +201,10 @@ func (s Service) generateMerkle() error {
 	s.minedBlock.mut.Lock()
 	defer s.minedBlock.mut.Unlock()
 	for _, statechainBlock := range s.minedBlock.StatechainBlocksMap {
+		if s.props.Context.Err() != nil {
+			return s.props.Context.Err()
+		}
+
 		if statechainBlock == nil {
 			return errors.New("nil block")
 		}
@@ -210,10 +217,6 @@ func (s Service) generateMerkle() error {
 	}
 
 	tree, err := merkle.BuildFromObjects(list, merkle.StatechainBlocksKindStr)
-	//tree, err := merkle.New(&merkle.TreeProps{
-	//Hashes: hashes,
-	//Kind:   merkle.StatechainBlocksKindStr,
-	//})
 	if err != nil {
 		return err
 	}
@@ -296,9 +299,7 @@ func (s Service) buildNextStates(imageHash string, transactions []*statechain.Tr
 		return err
 	}
 
-	log.Printf("[miner] gathered prev state block for image hash %s", imageHash)
-
-	// TODO: upload a genesis block if nil
+	// TODO: upload a genesis block if nil?
 	if prevStateBlock == nil {
 		log.Printf("[miner] prev state block is nil for image hash %s", imageHash)
 		return errors.New("could not find most recent state block")
@@ -322,23 +323,24 @@ func (s Service) buildNextStates(imageHash string, transactions []*statechain.Tr
 	}
 	// note: prepend
 	diffs = append([]*statechain.Diff{diff}, diffs...)
-	log.Printf("[miner] diffs %v", len(diffs))
 
 	head := prevStateBlock
 	for head.Props().BlockNumber != mainchain.GenesisBlock.Props().BlockNumber {
+		if s.props.Context.Err() != nil {
+			return s.props.Context.Err()
+		}
+
 		prevStateCID, err := p2p.GetCIDByHash(head.Props().PrevBlockHash)
 		if err != nil {
 			log.Printf("[miner] error getting cid by hash for %s for image hash %s", head.Props().PrevBlockHash, imageHash)
 			return err
 		}
-		log.Printf("[miner] got prev state cid for image hash %s", imageHash)
 
 		prevStateBlock, err := s.props.P2P.GetStatechainBlock(prevStateCID)
 		if err != nil {
 			return err
 		}
 		head = prevStateBlock
-		log.Printf("[miner] set head to prev state block for image hash %s", imageHash)
 
 		diffCID, err := p2p.GetCIDByHash(prevStateBlock.Props().StatePrevDiffHash)
 		if err != nil {
@@ -365,7 +367,6 @@ func (s Service) buildNextStates(imageHash string, transactions []*statechain.Tr
 		log.Printf("[miner] error creating tmp state file %s", err)
 		return err
 	}
-	log.Printf("[miner] tmp state file %s", tmpStateFile.Name())
 	defer os.Remove(tmpStateFile.Name()) // clean up
 
 	if _, err := tmpStateFile.Write([]byte(genesisState)); err != nil {
@@ -387,12 +388,15 @@ func (s Service) buildNextStates(imageHash string, transactions []*statechain.Tr
 	}
 
 	for i, diff := range diffs {
+		if s.props.Context.Err() != nil {
+			return s.props.Context.Err()
+		}
+
 		tmpPatchFile, err := ioutil.TempFile("", fmt.Sprintf("%s/%v/patch.%d.txt", imageHash, ts, i))
 		if err != nil {
 			log.Printf("[miner] error creating tmp patch file %s", err)
 			return err
 		}
-		log.Printf("[miner] tmp patch file %s", tmpPatchFile.Name())
 		defer os.Remove(tmpPatchFile.Name()) // clean up
 
 		if _, err := tmpPatchFile.Write([]byte(diff.Props().Data)); err != nil {
@@ -420,7 +424,6 @@ func (s Service) buildNextStates(imageHash string, transactions []*statechain.Tr
 		return err
 	}
 
-	log.Printf("[miner] state for image hash %s\nstate: %s", imageHash, string(state))
 	headStateFileName := tmpStateFile.Name()
 	runningBlockNumber, err := hexutil.DecodeUint64(prevStateBlock.Props().BlockNumber)
 	if err != nil {
@@ -430,6 +433,10 @@ func (s Service) buildNextStates(imageHash string, transactions []*statechain.Tr
 
 	// apply state to container and start running transactions
 	for i, tx := range transactions {
+		if s.props.Context.Err() != nil {
+			return s.props.Context.Err()
+		}
+
 		if tx == nil {
 			log.Printf("[miner] tx is nil for image hash %s", imageHash)
 			return errors.New("nil tx")
@@ -538,13 +545,13 @@ func (s Service) buildNextStates(imageHash string, transactions []*statechain.Tr
 
 	// write to the mined block
 	s.minedBlock.mut.Lock()
+	defer s.minedBlock.mut.Unlock()
 	// note: they should all have same length
 	for i := 0; i < len(newDiffs); i++ {
 		s.minedBlock.DiffsMap[*newDiffs[i].Props().DiffHash] = newDiffs[i]
 		s.minedBlock.TransactionsMap[*newTxs[i].Props().TxHash] = newTxs[i]
 		s.minedBlock.StatechainBlocksMap[*newStatechainBlocks[i].Props().BlockHash] = newStatechainBlocks[i]
 	}
-	s.minedBlock.mut.Unlock()
 
 	return nil
 }
